# server/openapi.yaml
openapi: 3.1.0
info:
  title: QFree API
  version: 1.0.0
  description: |
    Smart healthcare queue management and real-time waiting-time information.

    **Auth:** send `Authorization: Bearer <accessToken>` (15 min). Renew with `POST /auth/refresh`,
    which reads the httpOnly `qf_rt` cookie and rotates it.

    **Errors:** `{ "error": { "code": "QUEUE_FULL", "message": "…", "details": … } }`.

    **Real-time:** Socket.IO on the same origin. Handshake `auth: { token }` (optional).
    Client → server: `queue:watch`, `queue:unwatch`, `queue:watch-staff`, `queue:unwatch-staff`, `org:watch` (arg: id, ack `{ok, error?}`).
    Server → client: `queue:snapshot` (QueueSnapshot), `queue:entry` (EntryView), `queue:staff` (StaffView), `org:queue`, `notification:new`.
servers:
  - url: /api
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: qf_rt }
  parameters:
    id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    pageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    q: { name: q, in: query, schema: { type: string, maxLength: 100 } }
    from: { name: from, in: query, schema: { type: string, format: date }, description: Default 29 days ago }
    to: { name: to, in: query, schema: { type: string, format: date }, description: Default today }
    bucket: { name: bucket, in: query, schema: { type: string, enum: [day, week, month], default: day } }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NoContent: { description: No content }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, QUEUE_FULL, QUEUE_CLOSED, QUEUE_PAUSED, ALREADY_IN_QUEUE, OUTSIDE_HOURS, DOCTOR_UNAVAILABLE, TOO_MANY_ACTIVE_QUEUES, RATE_LIMITED] }
            message: { type: string }
            details: {}
    Paged:
      type: object
      properties:
        items: { type: array, items: {} }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }
...
    Analytics:
      type: object
      properties:
        summary: { type: object, description: 'total, served, cancelled, noShow, skipped, avgWaitMinutes, p90WaitMinutes, avgConsultMinutes, completionRate, utilizationPercent' }
        volume: { type: array, items: { type: object } }
        peakHours: { type: array, items: { type: object, properties: { hour: { type: integer }, joins: { type: integer } } } }
        weekdays: { type: array, items: { type: object } }
        perQueue: { type: array, items: { type: object } }
    Hours:
      type: object
      properties:
        dayOfWeek: { type: integer, minimum: 0, maximum: 6 }
        openTime: { type: string, example: '09:00' }
        closeTime: { type: string, example: '18:00' }
        isClosed: { type: boolean }
security:
  - bearer: []
paths:
...
  /organizations:
    get:
      tags: [Discovery]
      security: []
      parameters:
        - $ref: '#/components/parameters/q'
        - { name: type, in: query, schema: { type: string, enum: [CLINIC, LABORATORY, HOSPITAL, DIAGNOSTIC_CENTER] } }
        - { name: city, in: query, schema: { type: string } }
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Paged' } } } } }
    post:
      tags: [Organizations]
      summary: Create an organization (ORG_ADMIN, or DOCTOR for their own practice); caller becomes OWNER
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name, type, address, city], properties: { name: { type: string }, type: { type: string }, address: { type: string }, city: { type: string }, timezone: { type: string, default: Asia/Kolkata }, hours: { type: array, items: { $ref: '#/components/schemas/Hours' } } } } } } }
      responses: { '201': { description: Created }, '403': { $ref: '#/components/responses/Error' } }
...
  /organizations/{id}:
    get: { tags: [Discovery], security: [], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: Profile, hours, doctors, services, live queues }, '404': { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Organizations], summary: Update profile (OWNER/MANAGER), parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK }, '403': { $ref: '#/components/responses/Error' } } }
...
  /organizations/{id}/manage:
    get: { tags: [Organizations], summary: Full management view (staff only), parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK } } }
  /organizations/{id}/hours:
    put: { tags: [Organizations], parameters: [{ $ref: '#/components/parameters/id' }], requestBody: { required: true, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Hours' } } } } }, responses: { '200': { description: OK } } }
  /organizations/{id}/staff:
    post: { tags: [Organizations], summary: Link an existing staff account or create one, parameters: [{ $ref: '#/components/parameters/id' }], responses: { '201': { description: Created } } }
  /organizations/{id}/staff/{memberId}:
    delete: { tags: [Organizations], parameters: [{ $ref: '#/components/parameters/id' }, { name: memberId, in: path, required: true, schema: { type: string, format: uuid } }], responses: { '204': { $ref: '#/components/responses/NoContent' } } }
  /organizations/{id}/doctors:
    post: { tags: [Organizations], summary: Link an existing doctor or create a doctor account, parameters: [{ $ref: '#/components/parameters/id' }], responses: { '201': { description: Created } } }
  /organizations/{id}/doctors/{doctorId}:
    delete: { tags: [Organizations], parameters: [{ $ref: '#/components/parameters/id' }, { name: doctorId, in: path, required: true, schema: { type: string, format: uuid } }], responses: { '204': { $ref: '#/components/responses/NoContent' } } }
  /organizations/{id}/services:
    post: { tags: [Organizations], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '201': { description: Created } } }
  /organizations/{id}/services/{serviceId}:
    patch: { tags: [Organizations], parameters: [{ $ref: '#/components/parameters/id' }, { name: serviceId, in: path, required: true, schema: { type: string, format: uuid } }], responses: { '200': { description: OK } } }
    delete: { tags: [Organizations], summary: Soft-delete, parameters: [{ $ref: '#/components/parameters/id' }, { name: serviceId, in: path, required: true, schema: { type: string, format: uuid } }], responses: { '204': { $ref: '#/components/responses/NoContent' } } }

  # ───────── Notifications ─────────
...
  /analytics/organizations/{id}:
    get: { tags: [Analytics], parameters: [{ $ref: '#/components/parameters/id' }, { $ref: '#/components/parameters/from' }, { $ref: '#/components/parameters/to' }, { $ref: '#/components/parameters/bucket' }], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Analytics' } } } } } }
...
  /admin/organizations/{id}:
    patch: { tags: [Admin], parameters: [{ $ref: '#/components/parameters/id' }], requestBody: { content: { application/json: { schema: { type: object, properties: { isVerified: { type: boolean }, isActive: { type: boolean } } } } } }, responses: { '200': { description: OK } } }
...